move `conf` directory to `setup` directory
